requirements.txt
torch>=2.2.0
transformers>=4.45.0
datasets>=3.0.0
peft>=0.13.0
trl>=0.12.0
accelerate>=1.0.0
bitsandbytes>=0.44.0
safetensors>=0.4.0
sentencepiece>=0.2.0
protobuf>=5.28.0
scipy>=1.14.0
pyyaml>=6.0.0
boto3>=1.35.0
huggingface-hub>=0.26.0
tokenizers>=0.20.0
evaluate>=0.4.0
scikit-learn>=1.5.0
tensorboard>=2.18.0
wandb>=0.18.0
